fix: move fragment error handling to utils instead
